Fix FloodedRiceCalculator and FloodedRiceSeasonCalculator
Removed calculation of FloodedRiceSeasonCalculator
for general (or I would guess main-season crop)
as no information was inputted for this from the
frontend. This was causing the calculator to fail
when trying to calculate the general season.
Fixed FloodedRiceSeasonCalculator by addressing
incorrect access of information for certain parameters
where module.parameter was replaced with
module.parent.parameter in order to actually
access the necessary information. (area for example)