Add necessary logic for activity completion
This commit adds the necessary logic for activity completion. It includes
the following changes:
- adding completion percentage to the activity model
- adding a function to calculate the completion percentage in utilities.py
- adding a call to the function in the views.py file, upon module PUT
- adding a migration file to update the database schema

An IN PROGRESS status was also added to the statuses on the DB.
If an activity is marked as IN PROGRESS, the completion percentage
is > 0 and < 1. If it is READY completion_percentage = 1, and
it is EMPTY if completion_percentage = 0.