Fix AnnualCropCalculator
The AnnualCropCalculator was not working as intended.
The issue lied with the evaluation of the organic_input_flt
variable. It was being set based on the organic_input_
type_start variable. This worked in the case in which
we were not in LUC, but when we were in LUC, the
variable might have been none. Therefore it has to
be calculated in each possible scenario based on
whether we are in start-w or wo