Remove get_grassland_soc from Perennial, Flooded
The function returned an object with fi, fmg, flu
and no SOC value, however this function was used
for the calculation of the SOC value in the start
scenario. Didn't make much sense, as then it was
sending this to the model which couldn't use it.
Original code is still there however commented out,
see with Claudio once he is back.