Add Activity save logic
If an Activity with the same name is trying to be
created in the Project, a number is appended in
brackets. Request from the front-end due to UE
necessities as this check is at the end of the
process. In the future maybe add an EndPoint for
them to call to see if an activity exists with
that name in the project.