Fix OrganicSoilCalculator dry_matter
dry_matter was retrieved regardless of whether
fire was used or not. As a check on whether fire
is used I considered whether a fire_use_type has
been selected. If it has not been selected, then
fire is not used and dry_matter should not be
retrieved. None is sent to the model as the
model only calculates if fire is used.