Fix OrganicSoil and LUC issue in Activity Copy
When copying an Activity an error occurred for
OrganicSoil as there is to be a 1 to 1 relationship
between OrganicSoil and module, same as for
LandUseChange. This was fixed by adding the same
logic as for LandUseChange where a copy of the
OrganicSoil is created and linked to the module.
However, there was an issue with the approach
previously, as LUC was also found in find_modules
therefore two copies of LUC were created. This was
fixed by adding a check with the continue statement

CHECK THE FUNCTIONING CLAUDIO although it all seems
to be working fine