Fix forest_management.py check for disturbances
This is a commit made on the mathematical model
to fix an issue of communication between the back
end and the front end. Absence of values should be
seen as a None, not a 0. When in future the issue
is resolved, the code should be reverted to
the original state.